Updated to version V1.1
Uploaded the new version of the code, which now also contains the new optimization method based on surrogateopt, described in the EJC version of the paper. Since this makes the fmincon based variant somewhat irrelevant, but the fmincon code is still needed to generate the figures for the ECC version of the paper, it was put in the fmincon_legacy subfolder. Perhaps in the future, this folder will be deleted, as the ECC version will be replaced by the EJC version. The code in the fmincon_legacy folder is self-contained, and needs no additional files from the main directory.