Preparations for V2.0
Most of the code was rewritten, so as to be far less confusing. It was also made CORA-friendly, so that it is now almost ready to be implemented as is (several changes will be made for CORA, for example it will be adapted to constrained zonotopes instead of zonotopes, as this is more general; also, some new methods may appear in CORA at a later time).

Due to the adaptation, some methods such as the vertex enumeration and the polyhedral norm algorithms were optimized, so that they actually give better runtime results than they used to give in V1.1. The change is significant, but the shape of the curves remains similar (i.e., the runtimes have been cut by approximately half in every case). This is, however, not currently apparent, since the latest simulations were run in summer, and thus exhibit slower runtimes as simulations run in winter, using the same algorithms (at least that is my best guess as to why suddenly the same algorithms output significantly slower runtimes).

Some algorithm-names were also modified, again for the sake of CORA-friendliness.